requirements.txt
langchain>=0.1.6
langchain-community>=0.0.28
langchain-openai>=0.0.5
python-dotenv>=1.0.1
psycopg2-binary>=2.9.9
supabase>=2.3.1
ipython>=8.22.2
openai>=1.12.0
python-dotenv>=1.0.1
typing-extensions>=4.9.0
pydantic>=2.6.3
pinecone-client>=3.0.0 